requirements.txt
fastapi==0.70.0
azure-cosmos==4.2.0
uvicorn==0.15.0